LayoutTests/media/modern-media-controls/volume-support/volume-support-media-api-mute.html

<script src="../../../resources/js-test-pre.js"></script>
<script src="../resources/media-controls-loader.js" type="text/javascript"></script>
<body>
<video src="../../content/test.mp4" style="width: 320px; height: 240px;" controls></video>
<div id="shadow"></div>
<script type="text/javascript">

window.jsTestIsAsync = true;

description("Testing the <code>VolumeSupport</code> behavior by using the media API to mute.");

const shadowRoot = document.querySelector("div#shadow").attachShadow({ mode: "open" });
const media = document.querySelector("video");
const mediaController = createControls(shadowRoot, media, null);

shouldBe("mediaController.controls.volumeSlider.value", "1");

media.addEventListener("volumechange", () => {
    shouldBe("mediaController.controls.volumeSlider.value", "0");
    debug("");
    shadowRoot.host.remove();
    media.remove();
    finishJSTest();
});

media.muted = true;

</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
